apps/api/pyproject.toml
[project]
name = "dashboard-api"
version = "0.1.0"
description = "FastAPI backend for GitHub Issues Dashboard"
readme = "README.md"
authors = [
    { name = "dam00042", email = "dam00042@red.ujaen.es" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "pynacl>=1.5.0",
    "pydantic-settings>=2.13.1",
    "tzdata>=2025.2",
    "uvicorn[standard]>=0.42.0",
]

[project.scripts]
dashboard-api = "dashboard_api.__main__:main"
dashboard-api-dev = "dashboard_api.dev_server:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
    "--cov=dashboard_api",
    "--cov-report=term-missing",
]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D203", "D213"]

